Internal use. Converts a single outcome learner into a multivariate outcome learner that predicts a matrix of predictions obtained by evaluating the single outcome learner at each possible value of variable_stratify.

Internal use. Converts a single outcome learner into a multivariate outcome learner that predicts a matrix of predictions obtained by evaluating the single outcome learner at each possible value of variable_stratify.

Internal use. Converts a single outcome learner into a multivariate outcome learner that predicts a matrix of predictions obtained by evaluating the single outcome learner at each possible value of variable_stratify.

Super classes

sl3::Lrnr_base -> sl3::Lrnr_stratified -> Lrnr_stratified_multivariate

Methods

Public methods

Lrnr_stratified_multivariate$new()

Lrnr_stratified_multivariate$clone()

Inherited methods

sl3::Lrnr_base$assert_trained()

sl3::Lrnr_base$base_chain()

sl3::Lrnr_base$base_predict()

sl3::Lrnr_base$base_train()

sl3::Lrnr_base$chain()

sl3::Lrnr_base$custom_chain()

sl3::Lrnr_base$get_outcome_range()

sl3::Lrnr_base$get_outcome_type()

sl3::Lrnr_base$predict()

sl3::Lrnr_base$predict_fold()

sl3::Lrnr_base$print()

sl3::Lrnr_base$process_formula()

sl3::Lrnr_base$reparameterize()

sl3::Lrnr_base$retrain()

sl3::Lrnr_base$sample()

sl3::Lrnr_base$set_train()

sl3::Lrnr_base$subset_covariates()

sl3::Lrnr_base$train()

sl3::Lrnr_base$train_sublearners()

Method new()

Usage

Lrnr_stratified_multivariate$new(learner, variable_stratify, ...)

Method clone()

The objects of this class are cloneable with this method.

Usage

Lrnr_stratified_multivariate$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.